Major rewrite of the site structure.
The new model is to keep track of individual Turns, recording what happens in each turn and building up the game state at a given turn on the fly. This allows us to go back in time to a previous turn and re-play the simulation if we find a mistake (which seems likely to happen). It doesn't record the entire board state, however, so this doesn't allow replays of the game.